+++ src/device-wallet/messages.go
@@ -7,7 +7,7 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/skycoin/skycoin/src/cipher"
 
-	messages "github.com/skycoin/hardware-wallet-go/src/device-wallet/messages/go"
+	messages "github.com/skycoin/hardware-wallet-protob/go"
 )
 
 // MessageCancel prepare Cancel request
@@ -88,6 +88,36 @@ func MessageAddressGen(addressN, startIndex int, confirmAddress bool) ([][64]byt
 	}
 
 	chunks := makeSkyWalletMessage(data, messages.MessageType_MessageType_SkycoinAddress)
+	return chunks, nil
+}
+
+// MessageDeviceGetRawEntropy prepare GetEntropy request
+func MessageDeviceGetRawEntropy(entropyBytes uint32) ([][64]byte, error) {
+	getEntropy := &messages.GetRawEntropy{
+		Size_: &entropyBytes,
+	}
+
+	data, err := proto.Marshal(getEntropy)
+	if err != nil {
+		return [][64]byte{}, err
+	}
+
+	chunks := makeSkyWalletMessage(data, messages.MessageType_MessageType_GetRawEntropy)
+	return chunks, nil
+}
+
+// MessageDeviceGetMixedEntropy prepare GetMixedEntropy request
+func MessageDeviceGetMixedEntropy(entropyBytes uint32) ([][64]byte, error) {
+	getEntropy := &messages.GetMixedEntropy{
+		Size_: &entropyBytes,
+	}
+
+	data, err := proto.Marshal(getEntropy)
+	if err != nil {
+		return [][64]byte{}, err
+	}
+
+	chunks := makeSkyWalletMessage(data, messages.MessageType_MessageType_GetMixedEntropy)
 	return chunks, nil
 }